fixing region modification

+++ src/Region.h
@@ -10,7 +10,7 @@ class Region{
   public: 
     Region();
     Region(Cell startPoint, Cell endPoint);
-    std::vector<int> regionToVector(); // return [3,3,3] for a 3x3 grid
+    std::vector<uint8_t> regionToVector(); // return [3,3,3] for a 3x3 grid
     bool modify(Region * region);
     uint8_t numberOfSteps();
     std::vector<Cell> cells;
